Add pcli command to claim deposit refunds

+++ pcli/src/command/validator.rs
@@ -103,7 +103,7 @@ impl ValidatorCmd {
                     File::open(file).with_context(|| format!("cannot open file {:?}", file))?;
                 let new_validator: Validator = serde_json::from_reader(definition_file)
                     .map_err(|_| anyhow::anyhow!("Unable to parse validator definition"))?;
-                let fee = Fee::from_staking_token_amount((*fee as u64).into());
+                let fee = Fee::from_staking_token_amount((*fee).into());
 
                 // Sign the validator definition with the wallet's spend key.
                 let protobuf_serialized: ProtoValidator = new_validator.clone().into();
@@ -158,7 +158,7 @@ impl ValidatorCmd {
                 let vote = ValidatorVote { body, auth_sig };
 
                 // Construct a new transaction and include the validator definition.
-                let fee = Fee::from_staking_token_amount((*fee as u64).into());
+                let fee = Fee::from_staking_token_amount((*fee).into());
                 let plan = plan::validator_vote(
                     &app.fvk,
                     app.view.as_mut().unwrap(),
